backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-socketio[client]
